requirements.txt
openai>=1.12.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
tiktoken>=0.5.0
pytest>=7.4.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
apscheduler>=3.10.4
